requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
requests>=2.31.0
websockets>=12.0
cryptography>=41.0.0
Pillow>=10.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
